Inline providers into services
