--- common/src/main/java/com/wynntils/services/mapdata/MapDataProviders.java
@@ -1,119 +1,0 @@
-/*
- * Copyright © Wynntils 2023-2024.
- * This file is released under LGPLv3. See LICENSE for full license details.
- */
-package com.wynntils.services.mapdata;
-
-import com.wynntils.core.WynntilsMod;
-import com.wynntils.core.components.Services;
-import com.wynntils.services.mapdata.attributes.type.MapIcon;
-import com.wynntils.services.mapdata.providers.MapDataProvider;
-import com.wynntils.services.mapdata.providers.builtin.BuiltInProvider;
-import com.wynntils.services.mapdata.providers.builtin.CategoriesProvider;
-import com.wynntils.services.mapdata.providers.builtin.CharacterProvider;
-import com.wynntils.services.mapdata.providers.builtin.CombatListProvider;
-import com.wynntils.services.mapdata.providers.builtin.MapIconsProvider;
-import com.wynntils.services.mapdata.providers.builtin.PlaceListProvider;
-import com.wynntils.services.mapdata.providers.builtin.ServiceListProvider;
-import com.wynntils.services.mapdata.providers.builtin.WaypointsProvider;
-import com.wynntils.services.mapdata.providers.json.JsonProvider;
-import com.wynntils.services.mapdata.type.MapCategory;
-import com.wynntils.services.mapdata.type.MapFeature;
-import java.io.File;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.Map;
-import java.util.stream.Stream;
-
-public class MapDataProviders {
-    private static final MapDataProvider ONLINE_PLACEHOLDER_PROVIDER = new PlaceholderProvider();
-
-    private final LinkedList<String> providerOrder = new LinkedList<>();
-    private final Map<String, MapDataProvider> allProviders = new HashMap<>();
-
-    public MapDataProviders() {
-        createBuiltInProviders();
-    }
-
-    Stream<MapDataProvider> getProviders() {
-        return providerOrder.stream().map(allProviders::get);
-    }
-
-    // per-account, per-character or shared
-    // can be added just from disk, or downloaded from an url
-    public void createBundledProvider(String id, String filename) {
-        String completeId = "bundled:" + id;
-        JsonProvider provider = JsonProvider.loadBundledResource(completeId, filename);
-        registerProvider(completeId, provider);
-    }
-
-    public void createLocalProvider(String id, String filename) {
-        String completeId = "local:" + id;
-        JsonProvider provider = JsonProvider.loadLocalFile(completeId, new File(filename));
-        registerProvider(completeId, provider);
-    }
-
-    public void createOnlineProvider(String id, String url) {
-        String completeId = "online:" + id;
-        JsonProvider.loadOnlineResource(completeId, url, this::registerProvider);
-        // Register a dummy provider; this will be replaced once loading has finished
-        registerProvider(completeId, ONLINE_PLACEHOLDER_PROVIDER);
-    }
-
-    public void prioritizeProvider(String providerId) {
-        // This functionality should be replaced with a general reordering of
-        // providers from a GUI
-        if (providerOrder.remove(providerId)) {
-            // If it existed, put it back first
-            providerOrder.addFirst(providerId);
-        }
-    }
-
-    private void createBuiltInProviders() {
-        // Metadata
-        registerBuiltInProvider(new CategoriesProvider());
-        registerBuiltInProvider(new MapIconsProvider());
-
-        // Locations
-        registerBuiltInProvider(new ServiceListProvider());
-        registerBuiltInProvider(new CombatListProvider());
-        registerBuiltInProvider(new PlaceListProvider());
-        registerBuiltInProvider(new CharacterProvider());
-        registerBuiltInProvider(new WaypointsProvider());
-    }
-
-    private void registerBuiltInProvider(BuiltInProvider provider) {
-        registerProvider("built-in:" + provider.getProviderId(), provider);
-    }
-
-    private void registerProvider(String providerId, MapDataProvider provider) {
-        if (provider == null) {
-            WynntilsMod.warn("Provider missing for '" + providerId + "'");
-            return;
-        }
-        if (!allProviders.containsKey(providerId)) {
-            // It is not previously known, so add it first
-            providerOrder.addFirst(providerId);
-        }
-        // Add or update the provider
-        allProviders.put(providerId, provider);
-        Services.MapData.invalidateResolvedCache();
-    }
-
-    private static class PlaceholderProvider implements MapDataProvider {
-        @Override
-        public Stream<MapFeature> getFeatures() {
-            return Stream.empty();
-        }
-
-        @Override
-        public Stream<MapCategory> getCategories() {
-            return Stream.empty();
-        }
-
-        @Override
-        public Stream<MapIcon> getIcons() {
-            return Stream.empty();
-        }
-    }
-}
